Renamed use case groups
