+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/group-classifieds.js
@@ -5,6 +5,8 @@
 import { goodsOffer } from "./uc-goods-offer.js";
 import { goodsServiceSearch } from "./uc-goods-service-search.js";
 import { serviceOffer } from "./uc-service-offer.js";
+import { goodsTransportSearch } from "./uc-goods-transport-search.js";
+import { goodsTransportOffer } from "./uc-goods-transport-offer.js";
 
 /*
 TODO: create use case icons
@@ -18,6 +20,8 @@ export const classifiedsGroup = {
     goodsOffer: goodsOffer,
     goodsSearch: goodsServiceSearch,
     serviceOffer: serviceOffer,
-    // TODO: servicesDemand: {},
+    // TODO: serviceSearch: serviceSearch,
+    goodsTransportSearch: goodsTransportSearch,
+    goodsTransportOffer: goodsTransportOffer,
   },
 };
