+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/group-personal-mobility.js
