+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/group-real-estate.js
